% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateForPlot.R
\name{aggregateForPlot}
\alias{aggregateForPlot}
\title{aggregateForPlot}
\usage{
aggregateForPlot(plotdata,
dimension="region",regionPlotorder=NULL)
}
\arguments{
\item{plotdata}{An MAgPIE object}

\item{dimension}{by "region", "name" or "year"}

\item{regionPlotorder}{allows to change the order of the regions by giving
a vector with the order}
}
\value{
A MAgPIE object of the same dimensions as plotdata
}
\description{
Aggregates Data for an area chart. The value of all underlying
regions/years/timesteps are added to the value of the current
region/year/timestep.
}
\examples{
\dontrun{
data(population_magpie)
#no aggregation
scratch_plot(population_magpie[,,2])
#aggregation
scratch_plot(aggregateForPlot(population_magpie[,,2],regionPlotorder=c(10,1:9)))
}
}
\author{
Benjamin Bodirsky
}
