% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alloc_plot.R
\name{alloc_plot}
\alias{alloc_plot}
\title{alloc_plot}
\usage{
alloc_plot(land,title="",level="reg",print=TRUE,x="Year",
weight="Value/1000",fill="Data1",colors=NULL,legend_name="Land Type",
xlab=NULL,ylab="Area [billion ha]",
norm=FALSE,ncol=10,file=NULL,scale=4)
}
\arguments{
\item{land}{A MAgPIE object created by the land function of the magpie
library (or a MAgPIE object of the same structure)}

\item{title}{Diagram title}

\item{level}{Aggregation level/Facetting. "reg" for a regional, "glo" for a
global or "cell" for a cellular plot. Alternatively, you can also define
here the facetting as used in the ggplot2 library, e.g. "Year~Region", or
".~Data1"}

\item{print}{TRUE for direct printing. FALSE for the return of an object
which can be modified before printing.}

\item{x}{What should be printed on the x axis? (Year,Data1,Data2,...)}

\item{weight}{What should be stacked on top of each other? You can also use
calculations here such as "Value/1000" (default = Value/1000)}

\item{fill}{What should be distinguished by colors? (Year,Data1,Data2,...)}

\item{colors}{Which colors should be used (Default colors will be used for
colors=NULL)}

\item{legend_name}{Legend name, default name will be used for
legend_name=NULL}

\item{xlab}{x axis label, default name will be used for ylab=NULL}

\item{ylab}{y axis label, default name will be used for ylab=NULL}

\item{norm}{normalise land, FALSE or TRUE}

\item{ncol}{number of colums used for cellular plot (level="cell")}

\item{file}{a file name the output should be written to using ggsave}

\item{scale}{scale factor used in ggsave}
}
\description{
Land allocation plot. This plot was created to be able to nicely plot the
output of the land function which can be found in the magpie library.
}
\examples{

  \dontrun{
  alloc_plot(land=land(gdx,level="reg"),title="Total land (si0+nsi0 | global)",level="glo")
  alloc_plot(land=land(gdx,level="reg"),title="Total land (si0+nsi0 | regional)")
  } 
}
\seealso{
\code{\link{comparisonplot},\link{histoplot},\link{get_layout_parameters},\link{scratch_legend}}
}
\author{
Jan Philipp Dietrich, Benjamin Leon Bodirsky, Misko Stevanovic,
Florian Humpenoeder
}
