% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.ggplot.R
\name{as.ggplot}
\alias{as.ggplot}
\title{as.ggplot}
\usage{
as.ggplot(x,scenario="default",asDate=T,rev=F,useDimNames=FALSE)
}
\arguments{
\item{x}{MAgPIE object or list of MAgPIE objects. For a list of MAgPIE
objects the name of the list entry (has to be character) is used as scenario
name.}

\item{scenario}{Name of the scenario (only used if x is not a list) or
position of the entry in the third dimension that contains the scenario name
(only if the scenario name is stored in the third dimension) or NULL
(automatic detection)}

\item{asDate}{Format the years as date (TRUE) or keep as is (FALSE);
deprecated (only kept for compatibility with older scripts)}

\item{rev}{reverse legend order (TRUE) or not (FALSE)}

\item{useDimNames}{Use dim names of 3rd dimension instead Data1, Data2,
Data3, Data4; works only if x is a MAgPIE object (no list) and scenario=NULL}
}
\value{
Dataframe usable for ggplot
}
\description{
Converts a MAgPIE object or a list of MAgPIE objects into a dataframe usable
for ggplot
}
\examples{

\dontrun{
as.ggplot(croparea(gdx))}
\dontrun{
prices <- list()
prices[["Scenario1"]] <- prices("Scenario1.gdx",crops=c("tece","maiz"))
prices[["Scenario2"]] <- prices("Scenario2.gdx",crops=c("tece","maiz"))
as.ggplot(prices)}
}
\author{
Florian Humpenoeder, David Klein
}
