% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deco_plot.R
\name{deco_plot}
\alias{deco_plot}
\title{deco_plot}
\usage{
deco_plot(x, color = gray.colors(6))
}
\arguments{
\item{x}{magpie object with decomposition data (e.g. from the function magpie4:::deco)}

\item{color}{color vector}
}
\value{
a ggplot object
}
\description{
Plot for decomposition analysis.
Use function deco in library magpie4 to create the necessary object.
}
\author{
Benjamin Leon Bodirsky, Ina Neher
}
