% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layout_parameters.R
\name{get_layout_parameters}
\alias{get_layout_parameters}
\title{get_layout_parameters}
\usage{
get_layout_parameters(ngraphics, nrow=2, ncol=5, header=FALSE,
design=ifelse(nrow>ncol,"vertical","horizontal"), verbose=FALSE)
}
\arguments{
\item{ngraphics}{Number of Graphics which shall be plotted}

\item{nrow}{Number of graphics per row}

\item{ncol}{Number of graphics per column}

\item{header}{adds an additional plot windows at the top of the layout
matrix (can be used for titel and/or legend)}

\item{design}{Horizontal or vertical Page design}

\item{verbose}{decides, if additional information about the settings should
be reported or not}
}
\value{
Creates a list with basic design parameters.
return_data<-list(layout_matrix,page_width,page_height,graphics_per_page,n_pages,design,is_left,is_right,is_bottom,is_top)
layout_matrix returns a matrix which corresponds to the arrangement of
graphics on the page. Example: 1,2,3 4,5,6
}
\description{
Estimates the layout parameters when multiple plots shall be plotted on one
page
}
\examples{
\dontrun{
layout_parameters<-get_layout_parameters(15, nrow=2, ncol=5, design="horizontal", verbose=TRUE) 
pdf(file="output.pdf",width=layout_parameters$page_width,height=layout_parameters$page_height)
layout(layout_parameters$layout_matrix)
layout(layout_parameters[[1]])  
      for(i in 1:15) 
      {
         plot((1:10))
      }                                                            
dev.off()
      }
}
\seealso{
\code{\link{comparisonplot},\link{histoplot}}
}
\author{
Benjamin Bodirsky, Jan Philipp Dietrich
}
