% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg.reg.x.crop.R
\name{gg.reg.x.crop}
\alias{gg.reg.x.crop}
\title{gg.reg.x.crop}
\usage{
gg.reg.x.crop(..., ylab="", legend_name="", legend_labels=NULL,
legend_position="bottom", plot_title=NULL, palette=NULL, y_scales="fixed",
y_scale_type="log10", y_breaks=c(0.1,1,10), h_line=NULL, h_color="black",
h_type="dashed")
}
\arguments{
\item{...}{Requires a MAgPIE object(s). Ideally with regional dimension. If
there are more than one objects, they should have the same dimension. In the
case of plotting more than 8 \code{MAgPIE} objects, it is necessary to
define new color palette (see \code{palette} argument description).  Use the
\code{MAgPIE} dimensionality to specify which regions and production
activities you want to plot. For the latter, the optimal number is 6
activities per plot for (landscape) printing on a standard A4 paper.}

\item{ylab}{A string for the y-axis label name to be printed on the plot.
Default value is an empty string, i.e. no label is printed on the plot.}

\item{legend_name}{A string for the legend title to be printed on the plot.
Default value is an empty string, i.e. no title is printed on the plot.}

\item{legend_labels}{A vector of names for the legend labels. If nothing is
specified, the legend labels will be just the names of \code{MAgPIE} objects
that are given as the first arguments. See the example for the specific
usage.}

\item{legend_position}{A legend position on the plot. Default is "bottom".
Other options are "top", "right" and "left".}

\item{plot_title}{A string for the main title of the plot. Default value is
\code{NULL} which puts no title on the plot.}

\item{palette}{A vector of specific colors for the plot lines. If skipped,
the predefined palette will be used. At the moment, 8 colors are defined in
the palette, which means no more than 8 \code{MAgPIE} objects can be used
for plotting.}

\item{y_scales}{Decides if the y-axis scale is "fixed" (default) or "free"
for all the separate plots.}

\item{y_scale_type}{Defines the scaling type of the y-axis. Default type is
\code{log10}. Another option is \code{sqrt} for the square root scaling. For
the full scale y-axis, type anything.}

\item{y_breaks}{Break points on the y-scale (ticks). Default is \code{c(0.1,
1, 10)}.}

\item{h_line}{Puts a horizontal line to the plot. Default value is
\code{NULL}. It accepts a number as an argument which is the intercept on
the y-axis.}

\item{h_color}{A color for the horizontal line. The default color is
\code{"black"}.}

\item{h_type}{A line type for the horizontal line. The default type is
\code{"dashed"}.}
}
\description{
Creates a facet plot with regions as columns and production activities as
rows.
}
\details{
Requires a MAgPIE object(s). Ideally with regional dimension. If there are
more than one objects, they should have the same dimension. In the case of
plotting more than 8 \code{MAgPIE} objects, it is necessary to define new
color palette (see \code{palette} argument description).  Use the
\code{MAgPIE} dimensionality to specify which regions and production
activities you want to plot. For the latter, the optimal number is 6
activities per plot for (landscape) printing on a standard A4 paper format.
}
\examples{

  \dontrun{
  gg.reg.x.crop(magpie.reg.prod, magpie.supply, magpie.export,
   legend_labels=c("Regional\nProduction",
  "Supply","Export"), plot_title="Plot", palette=c("yellow","red","green"))
  
  gg.reg.x.crop(self.suff("../Some_magpie_run_in_the_output_folder/fulldata.gdx")
  , palette="magenta", 
  legend_labels="Self-Sufficiency\nRatios", y_breaks=c(0,1,10))	
  
  You can specify your magpie object as much as possible. If no scaling
  of y-axis is required, type for 
  example an empty string.
  gg.reg.x.crop(magpie_reg_prod[AFR,c(1:4),c("tece","begr")], y_scale_type="")
  } 
}
\author{
Miodrag Stevanovic
}
