% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gginput.R
\name{gginput}
\alias{gginput}
\title{gginput}
\usage{
gginput(data, scatter=NULL, mappings=NULL, na.rm=FALSE, verbose=TRUE)
}
\arguments{
\item{data}{MAgPIE object which provides the data for ggplot}

\item{scatter}{In the case you want to make a scatterplot you have to
specify here which data dimension of the MAgPIE object should be used for
the scatter plot (name of the dimension)}

\item{mappings}{A list of mappings used to create combinations of columns
which then can be used for aesthetics in ggplot}

\item{na.rm}{Boolean which decides whether NA values should be removed from
the data.}

\item{verbose}{Boolean which decides whether info messages should be printed to screen or not.}
}
\value{
Dataframe usable for ggplot
}
\description{
Converts a MAgPIE object into a dataframe usable for ggplot
}
\examples{
\dontrun{
library(ggplot2)
data(population_magpie)
ggplot(gginput(population_magpie,scatter="scenario"),aes(x=.value.A2,y=.value.B1,color=i)) 
+ geom_point() + geom_smooth(method="loess") + facet_wrap(~.spat1)
}
}
\author{
Jan Philipp Dietrich
}
