% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histoplot.R
\name{histoplot}
\alias{histoplot}
\title{histoplot}
\usage{
histoplot(measured_data, magpie_data, main="Histoplot", ylab="data",
xlim=NULL, ylim=NULL, file_type="normal_device",
output_folder="",legend=NULL, lowpass.iter = NULL, cex=1, regions=NULL, ...)
}
\arguments{
\item{measured_data}{A MAgPIE object (will be plotted as dots) or a list of
such objects with identical dimnames}

\item{magpie_data}{Another MAgPIE object with dimnames identical to
measured_data or a list of such MAgPIE objects (will be plotted as line).}

\item{main}{Title appering at the head of the page (either a string or a
list of string and quotes - see example)}

\item{ylab}{a title for the y axis (either a string or a list of string and
quotes)}

\item{xlim}{the x limits (x1, x2) of the plot. Note that x1 > x2 is allowed
and leads to a "reversed axis".}

\item{ylim}{Upper end of the plot scale. If no value given, the highest
value in the dataset is used.}

\item{file_type}{Data can be viewed either on the R-Device or printed as
pdf. Use "pdf", "normal device", or "none"}

\item{output_folder}{Outputfolder of pdf file}

\item{legend}{Legend for the description of the different lines in the plot}

\item{lowpass.iter}{Number of iterations a lowpass-filter should be applied
on observed data (a averaged line will be plotted in addition to the
observation points). lowpass.iter = NULL means, that no line should be
plotted.}

\item{cex}{A numerical value giving the amount by which plotting text and
symbols should be magnified relative to the default}

\item{regions}{Allows to choose the regions that should be plotted}

\item{...}{additional parameters, such as design, nrow, or ncol can be
forwarded to the function get_layout_parameters}
}
\value{
Compares two sets of historical data in the form of MAgPIE objects.
Dimensions are time and value. Regions and the name of the data columns have
to be the same if data shall be compared.
}
\description{
Compares two Magpie objects of the same structure on a linechart
}
\examples{
\dontrun{
data(population_magpie)
# Comparison requires two objects with same dimension names!
# Suppose the A2 dataset would be Magpie-Output, while the B1 would be the real data
magpie_output<-population_magpie[,,1]
dimnames(magpie_output)[[3]]<-c("Population")
real_data<-population_magpie[,,2]
dimnames(real_data)[[3]]<-c("Population")
 Note that only the timeslots where plotted where information on both datasets existed!
histoplot(measured_data=real_data,magpie_data=magpie_output,ylim=c(0,3000),
main=list(quote(alpha)," und ",quote(beta)))
}
}
\seealso{
\code{\link{comparisonplot},\link{scratch_plot},\link{get_layout_parameters}}
}
\author{
Benjamin Bodirsky, Jan Philipp Dietrich, Michael Krause
}
