% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/histoplot2.R
\name{histoplot2}
\alias{histoplot2}
\title{histoplot2}
\usage{
histoplot2(data_mag,data_hist,hist_start=1960,color="Scenario",
ylab="Value",title=NULL,facet_x="Region",facet_y=NULL,
labs=c("MAgPIE output","Historical Data"),text_size=12,scales="fixed",ncol=5,
xlim=NULL,ylim=NULL,breaks_x=waiver(),
breaks_y=waiver(),xaxis_angle=90,xlab="Year",
linewidth=1,pointwidth=3,col_data=NULL,col_hist=NULL,show_grid=TRUE,
axis_text_col="black",space="fixed",legend_data_ncol=1,
legend_hist_ncol=1,facet_style="paper",na.rm=FALSE,...)
}
\arguments{
\item{data_mag}{MAgPIE object or list of MAgPIE objects or dataframe of
MAgPIE object(s); the name of the list entries is used as scenario name in
the legend}

\item{data_hist}{list of MAgPIE objects with historial data; the name of the
list entries is used as dataset name in the legend}

\item{hist_start}{historical data will be shown starting from the year
specified in hist_start}

\item{color}{Dimension to be colored, default: "Scenario"}

\item{ylab}{y-axis text}

\item{title}{title appering at the top of the plot}

\item{facet_x}{x-axis facet, default: "Region"}

\item{facet_y}{y-axis facet, default: NULL}

\item{labs}{vector with legend titles for data_mag and data_hist}

\item{text_size}{text size of axis, legend and title}

\item{scales}{fixed (default), free, free_y or free_x; ?facetgrid for
details}

\item{ncol}{Number of columns used in facet_wrap function}

\item{xlim}{x axis limits; NULL or vector with limits}

\item{ylim}{y axis limits; NULL or vector with limits}

\item{breaks_x}{Vector of x-axis breaks, by default ggplot2 takes the
decision}

\item{breaks_y}{Vector of y-axis breaks, by default ggplot2 takes the
decision}

\item{xaxis_angle}{Angle of xaxis text in degree, default=90}

\item{xlab}{x-axis text}

\item{linewidth}{linewidth, default=1}

\item{pointwidth}{pointwidth, default=3}

\item{col_data}{vector of colors for data_mag, if NULL colors are chosen
automaticly}

\item{col_hist}{vector of colors for data_hist, if NULL colors are chosen
automaticly}

\item{show_grid}{show minor and major grid lines; FALSE (default) or TRUE}

\item{axis_text_col}{color of axis text and ticks; default value: "black"}

\item{space}{fixed (default), free, free_y or free_x; ?facetgrid for details}

\item{legend_data_ncol}{Number of columns used in legend for data_mag}

\item{legend_hist_ncol}{Number of columns used in legend for data_hist}

\item{facet_style}{style of facets, default or paper}

\item{na.rm}{Boolean deciding whether NA values should be filtered out of
the data or not.}

\item{...}{Further options passed to as.ggplot}
}
\value{
ggplot2 object
}
\description{
Function for plotting MAgPIE results and historical data with ggplot2
}
\examples{

\dontrun{
gdx <- list()
gdx[["Test"]] <- "fulldata.gdx"
mag <- mapply("*",lapply(gdx,emissions,level="glo",
type="co2_c",cumulative=FALSE),44/12,SIMPLIFY=FALSE)
hist <- lapply(getData(emissions,gdx=gdx,type="co2_c",
level="glo")[[1]][[2]][[1]],function(x) return(x*44/12))
histoplot2(data_mag = mag,data_hist = hist,ylab = expression(bold(MtCO[2]/yr)),legend_hist_ncol = 2)}

}
\seealso{
\code{\link{scratch_plot},\link{histoplot},\link{magpie2ggplot2},}
}
\author{
Florian Humpenoeder
}
