% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luplot-package.R
\docType{package}
\name{luplot-package}
\alias{luplot-package}
\alias{luplot}
\title{Landuse Plot Library}
\description{
Some useful functions to plot data such as a map plot function for MAgPIE
objects
}
\details{
\tabular{ll}{ Package: \tab luplot\cr Type: \tab Package\cr Version: \tab
3.11\cr Date: \tab 2017-02-06\cr License: \tab LGPL-3\cr LazyLoad: \tab
yes\cr }
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pik-piam/luplot}
}

}
\author{
Benjamin Bodirsky, Jan Philipp Dietrich, Michael Krause, Miodrag
Stevanovic, Florian Humpenoeder

Maintainer: Benjamin Bodirsky <bodirsky@pik-potsdam.de>
}
