% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magpie2ggplot.R
\name{magpie2ggplot}
\alias{magpie2ggplot}
\title{magpie2ggplot}
\usage{
magpie2ggplot(data,ylab="Value",title="",facet_reg=T,shape_data=T,stack=F,
fill="Data1",alpha=NULL, legend_data="Data",legend_scenario="Scenario",
text_size=12,hline=NULL,legend_position="right",scales="fixed",
ncol=5,...)
}
\arguments{
\item{data}{MAgPIE object or list of MAgPIE objects or dataframe of MAgPIE
object(s); the name of the list entries is used as scenario name in the
legend}

\item{ylab}{y-axis text}

\item{title}{title appering at the top of the plot}

\item{facet_reg}{facets for each region (TRUE) or each MAgPIE object in data
(FALSE)}

\item{shape_data}{shape data (TRUE) or shape scenario (FALSE)}

\item{stack}{dotted line plot (FALSE) or stacked area plot (TRUE); only used
for facet_reg=FALSE.}

\item{fill}{fill "Region" or "Data1" in stacked plots}

\item{alpha}{NULL, "Region" or "Data1" in stacked plots}

\item{legend_data}{title of the legend for the 3rd dimension of the MAgPIE
object; only used if the MAgPIE object has more than 3 dimensions}

\item{legend_scenario}{title of the legend for the scenarios in data}

\item{text_size}{text size of axis, legend and title}

\item{hline}{NULL or numeric. Adds a horizontal line at y=hline}

\item{legend_position}{right (default), left, top, bottom or none}

\item{scales}{fixed (default), free, free_y or free_x}

\item{ncol}{Number of columns used in facet_wrap function}

\item{...}{Further options passed to as.ggplot}
}
\value{
ggplot2 object representing one or more MAgPIE objects.
}
\description{
Function for plotting MAgPIE objects with ggplot
}
\examples{

\dontrun{
crop_area_reg <- list()
crop_area_reg[["scenario_name"]] <- croparea("fulldata.gdx")
magpie2ggplot(crop_area_reg,ylab="10^6 ha",title="Croparea")}  
\dontrun{
prices <- list()
prices[["Scenario1"]] <- prices("Scenario1.gdx",crops=c("tece","maiz"))
prices[["Scenario2"]] <- prices("Scenario2.gdx",crops=c("tece","maiz"))
magpie2ggplot(prices,ylab="US$/ton DM",title="Agricultural prices",
facet_reg=T,legend_title="Commodity")}
}
\seealso{
\code{\link{scratch_plot},\link{histoplot},}
}
\author{
Florian Humpenoeder
}
