% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magpie2ggplot2.R
\name{magpie2ggplot2}
\alias{magpie2ggplot2}
\title{magpie2ggplot2}
\usage{
magpie2ggplot2(data,scenario="default",ylab="Value",title=NULL,
xaxis="Year",yaxis="Value",facet_x="Region", facet_y=NULL,
geom="line",stack=F,color="Scenario",fill="Region",
shape=NULL,linetype=NULL,alpha=NULL,labs=NULL,stat=NULL,
text_size=11,hline=NULL,legend_position="right",
scales="fixed",ncol=5,xlim=NULL,ylim=NULL,file=NULL,
scale=NULL, breaks_x=waiver(),breaks_y=waiver(),
xaxis_angle=90,xlab="Year",linewidth=1,pointwidth=2,
color_pal=NULL,bar_width=NULL,show_grid=FALSE,
group="Region",axis_text_col="black",zoom=FALSE,
space="fixed",legend_nrow=NULL,legend_ncol=NULL,
facet_style="default",na.rm=FALSE,
stack_share=FALSE,point_position="identity",
normalize=NULL,...)
}
\arguments{
\item{data}{MAgPIE object or list of MAgPIE objects or dataframe of MAgPIE
object(s); the name of the list entries is used as scenario name in the
legend}

\item{scenario}{If the scenario name is stored in the third dimension:
position of the entry in the third dimension that contains the scenario
name. Default: "default"}

\item{ylab}{y-axis text}

\item{title}{title appering at the top of the plot}

\item{xaxis}{x-axis of the plot, default: "Year"}

\item{yaxis}{y-axis of the plot, default: "Value"}

\item{facet_x}{x-axis facet, default: "Region"}

\item{facet_y}{y-axis facet, default: NULL}

\item{geom}{"line","point","bar", or "area". If "none" only the ggplot base object is returned.}

\item{stack}{stacked bar plot (TRUE) or not (FALSE)}

\item{color}{Dimension to be colored, default: "Scenario"}

\item{fill}{Dimension to be filled in stacked plots, default: NULL}

\item{shape}{Dimension to be shaped, default: NULL}

\item{linetype}{Dimension to differ in linetypes, default: NULL}

\item{alpha}{Dimension to be transparent in stacked plots, default: NULL}

\item{labs}{vector with legend titles for color,fill,shape,alpha,linetype}

\item{stat}{adds statistics (e.g. "sum" or "mean"), default: NULL}

\item{text_size}{text size of axis, legend and title}

\item{hline}{NULL or MAgPIE objects. Adds a horizontal line.}

\item{legend_position}{right (default), left, top, bottom or none}

\item{scales}{fixed (default), free, free_y or free_x; ?facetgrid for
details}

\item{ncol}{Number of columns used in facet_wrap function}

\item{xlim}{x axis limits; NULL or vector with limits}

\item{ylim}{y axis limits; NULL or vector with limits}

\item{file}{File name the output should be written to using ggsave}

\item{scale}{scaling of ggplot2 object before saving to file}

\item{breaks_x}{Vector of x-axis breaks, by default ggplot2 takes the
decision}

\item{breaks_y}{Vector of y-axis breaks, by default ggplot2 takes the
decision}

\item{xaxis_angle}{Angle of xaxis text in degree, default=90}

\item{xlab}{x-axis text}

\item{linewidth}{linewidth, default=1}

\item{pointwidth}{pointwidth, default=3}

\item{color_pal}{vector of colors defining the color palette, if NULL colors
are chosen automaticly}

\item{bar_width}{width of bars in geom bar, default value: NULL}

\item{show_grid}{show minor and major grid lines; FALSE (default) or TRUE}

\item{group}{dimension used for grouping; default value: "Region"}

\item{axis_text_col}{color of axis text and ticks; default value: "black"}

\item{zoom}{TRUE zooms the plot according to xlim and ylim, FALSE omits
values not the range of xlim and ylim}

\item{space}{fixed (default), free, free_y or free_x; ?facetgrid for details}

\item{legend_nrow}{Number of rows used in legend}

\item{legend_ncol}{Number of columns used in legend}

\item{facet_style}{style of facets, default or paper}

\item{na.rm}{Boolean deciding whether NA values should be filtered out of
the data or not.}

\item{stack_share}{stacked bar plot showing shares (TRUE) or absolut values
(FALSE)}

\item{point_position}{position of points; "identity" or position_dodge(width
= 1)}

\item{normalize}{NULL (default) or year which should be used to normalize
the data}

\item{...}{Further options passed to as.ggplot}
}
\value{
ggplot2 object representing one or more MAgPIE objects.
}
\description{
Function for plotting MAgPIE objects with ggplot2
}
\examples{

\dontrun{
crop_area_reg <- list()
crop_area_reg[["scenario_name"]] <- croparea("fulldata.gdx")
magpie2ggplot2(crop_area_reg,ylab="10^6 ha",title="Croparea",stack=T,facet_x="Scenario",color=NULL)}
\dontrun{
prices <- list()
prices[["Scenario1"]] <- prices("Scenario1.gdx",crops=c("tece","maiz"))
prices[["Scenario2"]] <- prices("Scenario2.gdx",crops=c("tece","maiz"))
magpie2ggplot2(prices,ylab="US$/ton DM",title="Agricultural prices")}
}
\author{
Florian Humpenoeder, David Klein
}
