% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map2image.R
\name{map2image}
\alias{map2image}
\title{Plot MAgPIE object as map}
\usage{
map2image(x, type = "pdf", name = NULL, folder = ".", years =
getYears(x), data = getNames(x), onefile = TRUE, res = c(30,
15),main_add=NULL, ...)
}
\arguments{
\item{x}{MAgPIE object or file name of MAgPIE file}

\item{type}{output type: "screen", "pdf", or "png"}

\item{name}{name to be used in file name, if NULL either the MAgPIE object
file name is used or the file is named "map2image"}

\item{folder}{folder the output should be written to}

\item{years}{years to use}

\item{data}{data columns to use}

\item{onefile}{(only pdf) should be plotted in one or in separate files?}

\item{res}{output resolution c(x,y)}

\item{main_add}{additional map title}

\item{\dots}{Additional arguments passed to plotmap}
}
\description{
Function to plot MAgPIE objects as maps (without limitations of plotmap)
}
\examples{

#magmap("test.mz")
}
\seealso{
\code{\link{plotmap}}
}
\author{
Jan Philipp Dietrich
}
