% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mipAreaOverlapFirstNameFacetSecondNameStacked.R
\name{mipAreaOverlapFirstNameFacetSecondNameStacked}
\alias{mipAreaOverlapFirstNameFacetSecondNameStacked}
\title{mipAreaOverlapFirstNameFacetSecondNameStacked}
\usage{
mipAreaOverlapFirstNameFacetSecondNameStacked(x, alpha = 0.5)
}
\arguments{
\item{x}{A magpie or quitte object, or a list of magpie or quitte objects.}

\item{alpha}{alpha for each layer}
}
\value{
Stacked area plot
}
\description{
Plots an values in an area plot over time with scenarios in overlapping layers, the first name dimension as facet and the second name dimension stacked. Only datasets with not more than one region are supported.
}
\examples{

\dontrun{ 
library(moinput)
countries<-c("USA")
tmp<-collapseNames(calcOutput("FAOmassbalance",aggregate=FALSE)[,,"dm"])
tmp<-tmp[countries,,c("cassav_sp","potato")][,,c("food","feed")]
mipAreaOverlapFirstNameFacetSecondNameStacked(x=tmp)

}
}
\author{
Benjamin Leon Bodirsky
}
