% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nice_colors.R
\name{nice_colors}
\alias{nice_colors}
\title{nice_colors}
\usage{
nice_colors(style="contrast_area",saturation=1,alpha=1,value=1)
}
\arguments{
\item{style}{pre-defined Color-Scheme.  "contrast_area" returns
partial-complementary colors, with reduced saturation with rising number of
colors. This is useful for area-diagramms, where the focus of the eye shall
be forced to the bottom of the diagramm.}

\item{saturation}{saturation (0-1)}

\item{alpha}{Transparency (Attention, there are problems with certain
graphic devices or printers!}

\item{value}{Brightness. If brightness is part of the color-scheme, value is
a reduction factor of existing brightness.}
}
\value{
Vector of Hexa-Format of colors.
}
\description{
Nice Color Schemes
}
\examples{
\dontrun{
pie(1:12,col=nice_colors())
}
}
\author{
Benjamin Bodirsky
}
