% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVGsankeyNitrogen.R
\name{nitrogen_sankey}
\alias{nitrogen_sankey}
\title{nitrogen_sankey}
\usage{
nitrogen_sankey(x, svg_in, svg_out = "sankey.svg")
}
\arguments{
\item{x}{...}

\item{svg_in}{...}

\item{svg_out}{...}
}
\description{
...
}
\examples{
\dontrun{
...
}

}
