% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performancePlot.R
\name{performancePlot}
\alias{performancePlot}
\title{performancePlot}
\usage{
performancePlot(
  pdata,
  odata,
  fname = NULL,
  weight = NULL,
  measures = NULL,
  p_value = FALSE,
  qnt = NULL,
  w = 8.5,
  h = 11,
  mfrow = c(1, 1),
  f_siz = 0.6,
  na.rm = FALSE,
  drange = NULL,
  ...
)
}
\arguments{
\item{pdata}{Predicted data (e.g. MAgPIE output)}

\item{odata}{Observed data (e.g. FAO)}

\item{fname}{Name of the pdf output file. If the \code{fname} is
\code{NULL}, the pdf file will not be created, but the created plots will be
previewed in the R graphic device.}

\item{weight}{A matrix of weights. Must have the same row and column names
as the data objects. Can be unspecified in order to obtain the unweighted
mean.}

\item{measures}{A vector of efficiency measures that are wanted to be
calculated and reported on the plots. Currently available measures are:
\code{"Willmott", "Willmott refined", "Nash Sutcliffe", "RMSE", "MAE",
"Pearson","Kendall","Spearman"}}

\item{p_value}{p_value=TRUE,reports significance level of Pearson, Kendall
and Spearmann coefficients.}

\item{qnt}{A parameter for labeling the points on the plot. It will label
all the points that are greater than the percentile given with this
parameter. If \code{qnt=NULL} no point will be labeled. If \code{qnt="0\%"},
all labels are put in the plot. If the data has no names (row names), leave
\code{qnt=NULL}.}

\item{w}{Width of the pdf document.}

\item{h}{Height of the pdf document.}

\item{mfrow}{A vector of the form \code{c(nr,nc)}. Subsequent figures will
be drown in an \code{nr}-by-\code{nc} array on the device by rows.}

\item{f_siz}{The size of font for point labels and reported indices on the
plots.}

\item{na.rm}{If \code{TRUE}, removes all the \code{NA}, \code{NaN} or
\code{Inf} values in both corresponding sets.}

\item{drange}{The scale of the y-axis}

\item{...}{Additional arguments to be passed to the \code{plot()} function}
}
\description{
Creates scatter plots for comparison of two data sets and reports various
efficiency measures
}
\details{
The format of \code{pdata} and \code{odata} supported are \code{matrix},
\code{magpie} and \code{vector}. Comparison of one \code{matrix} and one
\code{magpie} object is allowed, but with the same row and column names. For
the \code{magpie} objects it compares just the first year in the \code{year}
dimension.  The comparison is performed columnwise.
}
\examples{

	\dontrun{
	performancePlot(magpie_prod,fao_prod,qnt="98\%",mfrow=c(3,2),fname="test",
	measures=c("Willmott","Pearson"))
	performancePlot(magpie_prod,fao_prod)
	performancePlot(magpie_prod, fao_prod, fname="MAgPIE_FAO_production",
	w=8.5, h=12, mfrow=c(3,2), f_siz=0.7, na.rm=TRUE, xlab=paste("MAgPIE",
	"production, mio.ton DM"),ylab=paste("FAO","production, mio.ton DM")
	,col="blue", drange=500)
	}

}
\seealso{
\code{\link{qualityMeasure}}
}
\author{
Misko Stevanovic,Xiaoxi Wang
}
