% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrHist2D.R
\name{plotCorrHist2D}
\alias{plotCorrHist2D}
\title{correlationDataSets}
\usage{
plotCorrHist2D(
  x,
  y,
  title = NULL,
  xlab = "x",
  ylab = "y",
  bins = 40,
  limx = NULL,
  limy = NULL,
  folder = NULL,
  file = "",
  breaks = waiver(),
  nrows = 2,
  ncols = 2,
  axisFont = 13,
  axisTitleFont = 13,
  TitleFontSize = 15,
  legendTitleFont = 12,
  legendTextFont = 10,
  statFont = 4,
  table = FALSE,
  stat = TRUE,
  palette = "RdYlBu",
  tag = NULL
)
}
\arguments{
\item{x}{First data set (x axis) (observed) as magpie object with one or more items in 3rd dimension}

\item{y}{Second data set (y axis) (predicted) as magpie object with one or more items in 3rd dimension}

\item{title}{title of plot}

\item{xlab}{x axis title}

\item{ylab}{y axis title}

\item{bins}{number of bins in histogram}

\item{limx}{limits y axis (if NULL, fitting limits are calculated within the function)}

\item{limy}{limits x axis (if NULL, fitting limits are calculated within the function)}

\item{folder}{path in which to save the plots. if "." is used, it is saved 
in the current working directory. If NULL, no plots are saved.}

\item{file}{name of file}

\item{breaks}{breaks of the legend. It can be a vector, waiver()=the ones from the transformation, NULL for no breaks}

\item{nrows}{number of rows in pdf file where plots are printed}

\item{ncols}{number of columns in pdf file where plots are printed}

\item{axisFont}{Font size of text of axis of the correlation plot}

\item{axisTitleFont}{Font size of title of axis of the correlation plot}

\item{TitleFontSize}{Font size of title of correlation plot}

\item{legendTitleFont}{Font size of the title of the legend}

\item{legendTextFont}{Font size of legend}

\item{statFont}{font size r2-MAE label}

\item{table}{Conditional to include table with statistics in the output. TRUE (includes it), FALSE (it doesn't)}

\item{stat}{Conditional to include R2 and MAE on the fiure. TRUE (includes it), FALSE (it doesn't)}

\item{palette}{palette selection for heatd maps based on the RColorBrewer library}

\item{tag}{for multiple items in the second and third dimensions of the magpie object, should the title include "year", "item", "year-item".}
}
\description{
Plots 2D density plots to evaluate the relationship between two datasets.
It also returns the R-square for the correlation
}
\examples{
\dontrun{
x <- plotCorrHist2D(x, y, folder = ".")
}

}
\author{
Edna Molina Bacca
}
