% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotcountrymap.R
\name{plotcountrymap}
\alias{plotcountrymap}
\title{Function to plot maps}
\usage{
plotcountrymap(x, hatching=FALSE, ...)
}
\arguments{
\item{x}{a magpie object with ISO3 country names, one year and one name
dimension}

\item{hatching}{if hatching is activated, the second name column will be
used for hatching (stripes)}

\item{...}{Further attributes which are passed on to the function
mapCountryData of the package rworldmap}
}
\description{
Plots maps of map objects or MAgPIE objects
}
\details{
This function is an easy application of the mapCountryData function for ISO3
country-level magpie objects
}
\examples{
\dontrun{
data(population_magpie)
test<-population_magpie
dimnames(test)[[1]]<-c("AFG","DEU","FRA","EGY","IND","IDN","RUS","CHN","USA","YEM")
plotcountrymap(test[,"y2005",1])
}
}
\seealso{
\code{\link{magpie2map}},\code{\link{plotmap}},\code{\link{plotmap2}}
}
\author{
Benjamin Leon Bodirsky
}
