% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmap2.R
\name{plotmap2}
\alias{plotmap2}
\title{Function to plot maps using ggplot2}
\usage{
plotmap2(data,file=NULL,title="World map",legend_range=NULL,
legendname="Cell share",lowcol="grey95",midcol="orange",
highcol="darkred",midpoint=0.5, facet_grid="Year~Data1", nrow=NULL, ncol=NULL, 
scale=2,breaks=T,labs=T,borders=TRUE,MAgPIE_regions=F, 
axis_text_col="black",legend_discrete=FALSE,legend_breaks=NULL,
show_percent=FALSE, sea=TRUE,land_colour="white",
legend_height=2,legend_width=NULL,text_size=12,
legend_position="right",facet_style="default",plot_height=10,plot_width=20)
}
\arguments{
\item{data}{MAgPIE object}

\item{file}{File name the output should be written to using ggsave}

\item{title}{Title of the map}

\item{legend_range}{Legend range. Vector with minimum and maximum.}

\item{legendname}{Name of the legend}

\item{lowcol}{color for low values}

\item{midcol}{color for medium values}

\item{highcol}{color for high values}

\item{midpoint}{Value indicating where the transformation from lowcol to
highcol takes place.}

\item{facet_grid}{facets formula used in ggplot2 with default "Year~Data1". Falls back to facet_wrap for "Data1".}

\item{nrow}{number rows for facet_wrap}

\item{ncol}{number columns for facet_wrap}

\item{scale}{scaling of ggplot2 object before saving to file}

\item{breaks}{TRUE or FALSE}

\item{labs}{TRUE or FALSE}

\item{borders}{TRUE or FALSE}

\item{MAgPIE_regions}{TRUE or FALSE, if TRUE data is ignored}

\item{axis_text_col}{color of axis text and ticks; default value: "black"}

\item{legend_discrete}{TRUE or FALSE, discrete or continous legend}

\item{legend_breaks}{NULL (automatically) or vector with breaks}

\item{show_percent}{TRUE or FALSE, percent values for discrete lengeds}

\item{sea}{TRUE or FALSE, blue sea or just white background}

\item{land_colour}{Background colour for all land area. Defaults to white.}

\item{legend_height}{NULL or legend height in cm}

\item{legend_width}{NULL or legend width in cm}

\item{text_size}{text size of axis, legend and title}

\item{legend_position}{right (default), left, top, bottom or none}

\item{facet_style}{style of facets, default or paper}

\item{plot_height}{plot height in cm}

\item{plot_width}{plot width in cm}
}
\description{
Plots maps MAgPIE objects
}
\details{
Makes use of ggplot2 map plotting capabilites
}
\examples{
#plotmap2(data)
#' #@importFrom  RColorBrewer
}
\seealso{
\code{\link{plotmap}}
}
\author{
Florian Humpenoeder
}
