% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmap_discrete.R
\name{plotmap_discrete}
\alias{plotmap_discrete}
\title{Function to plot maps with a discrete scale using ggplot2}
\usage{
plotmap_discrete(data,file=NULL,title="World map",
colours="YlOrRd",legend_breaks=NULL,legendname="share",
show_percent=TRUE,facet_grid="Year~Data1")
}
\arguments{
\item{data}{MAgPIE object}

\item{file}{File name the output should be written to using ggsave}

\item{title}{Title of the map}

\item{colours}{Determines the colour palette of the map. Can either be the
name of a colorbrewer scale (or the corresponding number, see the 'palette'
argument in \code{\link[ggplot2]{scale_fill_brewer}}, or a vector of
manually determined colours.}

\item{legend_breaks}{By default (NULL), the breaks are determined
automatically from the \code{data} object. Alternatively, manual breaks can
be set by providing a vector. Note that if \code{data} is continuous, breaks
have to be specified manually.}

\item{legendname}{Name of the legend}

\item{show_percent}{If TRUE, the percentage of cells in the corresponding
category is shown}

\item{facet_grid}{facets formula used in ggplot2 with default "Year~Data1"}
}
\description{
Plots maps of MAgPIE objects
}
\details{
Makes use of ggplot2 map plotting capabilites
}
\examples{
#plotmap2(data)
}
\seealso{
\code{\link{plotmap}},\code{\link{plotmap2}}
}
\author{
Florian Humpenoeder, Markus Bonsch
}
