% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNBudget.R
\name{plotNbudget}
\alias{plotNbudget}
\title{plotNbudget}
\usage{
plotNbudget(nitrogenbudget, type = "bar")
}
\arguments{
\item{nitrogenbudget}{A magpie object with a nitrogen budget}

\item{type}{"bar" or "area" plot}
}
\value{
Stacked bar or area plot
}
\description{
Plots Nitrogen Budgets for Cropland soils on regional or country levels.
}
\examples{

\dontrun{ 
library(moinput)
countries<-c("USA","CHN","NGA")
nitrogen_budget<-calcOutput("NitrogenBudget",aggregate = FALSE)
plotNbudget(nitrogenbudget=nitrogen_budget[countries,,],type="area")
plotNbudget(nitrogenbudget=colSums(nitrogen_budget),type="bar")

}
}
\author{
Benjamin Leon Bodirsky
}
