% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegionMapping.R
\name{plotRegionMapping}
\alias{plotRegionMapping}
\title{Function to plot maps}
\usage{
plotRegionMapping(file, col = NULL, ...)
}
\arguments{
\item{file}{a .csv file}

\item{col}{color}

\item{...}{Further attributes such as color palette and map name}
}
\description{
This function plots a world region map directly from .csv files
}
\examples{
\dontrun{
plotRegionMapping("../regionmapping.csv", col = NULL, mapTitle = "")
}
}
\seealso{
\code{\link{plotcountrymap}},\code{\link{plotmap2}}
}
\author{
Jan Philipp Dietrich, Ewerton Araujo
}
