% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRegions.R
\name{plotRegions}
\alias{plotRegions}
\title{plotRegions}
\usage{
plotRegions(mapping=NULL, cex=1, pt.cex=2)
}
\arguments{
\item{mapping}{A path to the mapping that provides iso to regional definition.}

\item{cex}{Controls the legend font size. Default is 1.}

\item{pt.cex}{Controls the legend color box size. Default is 2.}
}
\value{
Plots a regional map.
}
\description{
Maps world regions based on a provided mapping.
}
\details{
The function works only with the mappings as defined in the \code{inputdata/mappings/regional} folder
}
\examples{
\dontrun{
plotRegions(mapping="inputdata/mappings/regional/regionmappingMAgPIE.csv")
}

}
\author{
Mishko Stevanovic
}
