% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotregionscluster.R
\name{plotregionscluster}
\alias{plotregionscluster}
\title{plotregionscluster}
\usage{
plotregionscluster(x, cells = "magpiecell")
}
\arguments{
\item{x}{data frame with 59199 or 67420 cells in MAgPIE order
that contains regional and cluster information.
For 59199 cells: format of cluster dimension \code{region.cluster}
where \code{region} is the name of the region
and \code{cluster} is the number of the cluster.
For 67420 cells: cell dimension with coordinate and country name information
and format of cluster dimension \code{region.cluster}}

\item{cells}{"magpiecell" for 59199 cells and "lpjcell" for 67420 cells}
}
\description{
Plots world map showing world regions and cluster.
}
\examples{
x <- paste(rep(paste0("REG", 1:5), each = 12000), rep(1:25, each = 2400), sep = ".")[1:59199]
plotregionscluster(x)
}
