% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotspam.R
\name{plotspam}
\alias{plotspam}
\title{Plots Spam object as map}
\usage{
plotspam(spam,type="pdf",name="spamplot", folder=".")
}
\arguments{
\item{spam}{A spam relation object (either object itself or file name of a
spam object)}

\item{type}{output type: "screen", "pdf", or "png"}

\item{name}{name to be used in file name}

\item{folder}{folder the output should be written to}
}
\description{
Function to plot spam relation objects as maps
}
\examples{

\dontrun{plotspam("0.5-to-h100.spam")}
}
\seealso{
\code{\link{map2image}}
}
\author{
Jan Philipp Dietrich
}
