% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotUtilization.R
\name{plotUtilization}
\alias{plotUtilization}
\title{plotUtilization}
\usage{
plotUtilization(
  x,
  products = "kall",
  categories = c("Food", "Feed", "Processing", "Seed", "Supply Chain Loss", "Material",
    "Domestic Balanceflow")
)
}
\arguments{
\item{x}{A magpie or quitte object, or a list of magpie or quitte objects.}

\item{products}{Products}

\item{categories}{Categories}
}
\value{
Plot
}
\description{
Plot to visualize the utilisation of products for different purposes. Input should be an object derived from a mif file or similar.
}
\examples{

\dontrun{ 
require(mip)
require(moinput)
kall<-reportingnames(findset("kall"))
budget<-calcOutput("ValidDemand",nutrient="nr",aggregate = "GLO")
budget2<-reportingReverse(budget,select="Agricultural demand",namedims=3, endoftree=T)
plotUtilization(budget2,products = kall,categories=c("Food","Feed","Processing"))
budget2a<-budget2[,1:5]
budget2b<-budget2[,3:10]*1.1
getNames(budget2b,dim=1)<-"projection"
x<-list(colSums(budget2a),colSums(budget2b))
plotUtilization(x,products = kall,categories=c("Food","Feed","Processing"))
plotUtilization(x,products = findset("kall_agg1"),categories=c("Food","Feed","Processing"))
plotUtilization(x,products = findset("kall_agg2"),categories=c("Food","Feed","Processing"))
productsx=setdiff(findset("kall_agg1"),c("pasture","kres","foddr","bioenergycrops"))
plotUtilization(x,products = productsx,categories=c("Food","Feed","Processing"))
a<-calcOutput("ValidCroparea",detail=T)
a<-reportingReverse(a,select="Land Cover",namedims=3, endoftree=T)
plotUtilization(colSums(a),products = kall,categories = "Cropland")
}

}
\author{
Benjamin Leon Bodirsky
}
