% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityMeasure.R
\name{qualityMeasure}
\alias{qualityMeasure}
\title{qualityMeasure}
\usage{
qualityMeasure(pd, od, wt, measures=c("Willmott",
"Nash Sutcliffe"),p_value=FALSE)
}
\arguments{
\item{pd}{Predicted data. Format required: numeric vector.}

\item{od}{Observed data. Must be the same format and length as \code{pd}.}

\item{wt}{A vector of weights. The same length as \code{pd} and \code{od}.
Missing weights are interprated as all weights equal to 1.}

\item{measures}{A vector of supported statistical measures. Currently
available are: "Willmott", "Willmott refined", "Nash Sutcliffe", "RMSE",
"MAE", "Pearson","Kendall","Spearman".}

\item{p_value}{p_value=TRUE,reports significance level of Pearson, Kendall
and Spearmann coefficients.}
}
\value{
Returns a numeric vector with calculated measures. Names of vector
elements are the measures applied.
}
\description{
Function that returns several efficiency indices of model performance wrt
observed data
}
\examples{

\dontrun{
od <- rnorm(1000)
pd <- rnorm(1000)

qualityMeasure(pd,od)
qualityMeasure(pd,od,measures=c("RMSE","Pearson","Willmott refined"))
}
}
\seealso{
\code{\link{performancePlot}}
}
\author{
Misko Stevanovic,Xiaoxi Wang
}
