% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scratch_plot.R
\name{scratch_plot}
\alias{scratch_plot}
\title{scratch_plot}
\usage{
scratch_plot(plotdata,height=NULL,main=NULL,add=FALSE,ylab=NULL,
large=FALSE,pch=16,lty=rep(1:6,20),legend=TRUE,region_plotorder=NULL)
}
\arguments{
\item{plotdata}{An MAgPIE object to be visualized}

\item{height}{maximum height of the diagramm}

\item{main}{Title appering at the head of the page}

\item{add}{"region_name" makes an area-chart out of the MAgPIE object, with
the value of regions and names being plotted on top of each other.  "region"
makes an area-chart out of the MAgPIE object, with the value of only regions
being plotted on top of each other, while the name-slices are plotted in
front of each other with a transparency of 50 percent.}

\item{ylab}{y-axis text}

\item{large}{TRUE: font-size for postdocs :P}

\item{pch}{vector of dot-types, see par, only used for line-charts}

\item{lty}{vector of line-types, see par, only used for line-charts}

\item{legend}{legend can be switched on or of}

\item{region_plotorder}{reordering of regions, see \link{aggregate_for_plot}}
}
\value{
Aim of this function is to create a fast visualisation of an MAgPIE
object.
}
\description{
Function for a quick-and-dirty visualisation of a regional MAgPIE object.
}
\examples{
\dontrun{
  data(population_magpie)
  scratch_plot(population_magpie,main="Population Scenarios",ylab="Million People")
  scratch_plot(population_magpie,main="Population Scenarios",ylab="Million People", 
  add="region_name")
  scratch_plot(population_magpie,main="Population Scenarios",ylab="Million People",
   add="region_name",large=TRUE)  
  scratch_plot(population_magpie,main="Population Scenarios",ylab="Million People",
  add="region")    
}
}
\seealso{
\code{\link{comparisonplot},\link{histoplot},\link{get_layout_parameters},\link{scratch_legend}}
}
\author{
Benjamin Bodirsky
}
