% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showcolors.R
\name{showcolors}
\alias{showcolors}
\title{Creates bar plot to display given colors}
\usage{
showcolors(...)
}
\arguments{
\item{...}{One ore more strings or a vector of strings containing
hexadecimal color codes.}
}
\description{
For a quick translation of hexadecimal color codes into visible colors this
function simply creates a bar plot with the hexadecimal colors specified by
the user. It takes a vector of strings containing the hexadecimal color
codes, e.g. "#D0DD1E".
}
\examples{

\dontrun{showcolors("#D0DD1E")}
\dontrun{showcolors("#beefee")}
\dontrun{showcolors(c("#beefee","#D0DD1E"),"#FFFFFF")}
}
\author{
David Klein
}
