% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swoutput.R
\name{swoutput}
\alias{swoutput}
\title{swoutput}
\usage{
swoutput(stream,data,unit="unit",plot=TRUE,table=TRUE,scenario_name="default",
text_size=16,digits=0,plot_level=NULL,...)
}
\arguments{
\item{stream}{The \code{swStream} object to be modified}

\item{data}{MAgPIE object or list of MAgPIE objects. For a list of MAgPIE
objects the name of the list entry (has to be character) is used as scenario
name.}

\item{unit}{unit of data (Character). Used as ylab in
\code{\link[luplot]{magpie2ggplot}} and table caption in
\code{\link[xtable]{print.xtable}}}

\item{plot}{TRUE or FALSE. If TRUE, LaTeX code for plot
(\code{\link[luplot]{magpie2ggplot}}) is added}

\item{table}{TRUE or FALSE. If TRUE, LaTeX code for data table
(\code{\link[xtable]{print.xtable}}) is added}

\item{scenario_name}{Name of the scenario. Only used if data is not a list.}

\item{text_size}{Size of text used in \code{\link[luplot]{magpie2ggplot}}}

\item{digits}{Number of digits for rounding. Used in
\code{\link[xtable]{xtable}}}

\item{plot_level}{NULL for all (default), reg or glo}

\item{...}{Further options passed to \code{\link[luplot]{magpie2ggplot}} and
\code{\link[xtable]{print.xtable}}}
}
\value{
No return value
}
\description{
Creates LaTeX code for plot (\code{\link[luplot]{magpie2ggplot}}) and data
table (\code{\link[xtable]{print.xtable}}) based on MAgPIE objects and adds
it to a \code{"\linkS4class{swStream}"} object.
}
\details{
Creates preformatted plots and data tables based on MAgPIE objects. Format
is optimized for readability. Data tables with more than 12 columns are
rescaled for fitting on portrait pages.
}
\examples{

\dontrun{
sw<-swopen("croparea_reg.pdf")
swlatex(sw,"\\newpage")
swlatex(sw,"\\section{Croparea - regional}")
swoutput(sw,croparea(gdx),"mio. ha",scenario="test")
swclose(sw)}

}
\seealso{
\code{"\linkS4class{swStream}"},\code{\link{swopen}},\code{\link{swclose}},\code{\link{swR}},\code{\link{swtable}},\code{\link{swfigure}}
}
\author{
Florian Humpenoeder
}
