% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whereplot.R
\name{whereplot}
\alias{whereplot}
\title{whereplot}
\usage{
whereplot(x)
}
\arguments{
\item{x}{A logical statement with a magpie object on country resolution}
}
\value{
A map. Green is true, red is false, orange is true and false in the
same country, purple is NA, and striped indicates a country that includes
some NAs.
}
\description{
Plots the results of a logical question on a mpa
}
\examples{
\dontrun{
data(population_magpie)
 test<-population_magpie
 dimnames(test)[[1]]<-c("AFG","DEU","FRA","EGY","IND","IDN","RUS","CHN","USA","YEM")
 whereplot(test>500)
 }
}
\author{
Benjamin Leon Bodirsky
}
