% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luplot-package.R
\docType{data}
\name{wrld_simpl}
\alias{wrld_simpl}
\title{World countries' boundaries}
\description{
A shapefile dataframe containing up to date countriy borders (only used
internally by plotmap). The file is from the data provided in the maptools
package
}
\author{
Miodrag Stevanovic
}
