% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/luplot-package.R
\docType{data}
\name{wrld_simpl_df}
\alias{wrld_simpl_df}
\title{World countries' boundaries and MAgPIE mapping}
\description{
A dataframe containing up to date country borders and MAgPIE region mapping
(only used internally by plotmap2). The file is based on the wrld_simpl
dataset
}
\author{
Miodrag Stevanovic, Florian Humpenoeder
}
