% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotregioncluster.R
\name{plotregionscluster}
\alias{plotregionscluster}
\title{Plots world map with regions and clusters}
\usage{
plotregionscluster(x)
}
\arguments{
\item{x}{a vector with 59199 cells in MAgPIE order and format \code{region.cluster} where \code{region}
is the name of the region and \code{cluster} is the number of the cluster.}
}
\description{
Function to plot world map showing world regions and cluster.
}
\examples{
x <- paste(rep(paste0("REG", 1:5), each = 12000), rep(1:25, each = 2400), sep = ".")[1:59199]
  plotregionscluster(x)
}
\seealso{
\code{\link{plotspam}}
}
